+++ src/commands/mod.rs
@@ -1,6 +1,7 @@
 pub mod download;
 pub mod generate;
 pub mod git;
+pub mod install;
 pub mod ip;
 pub mod macos;
 pub mod push;